Fix Pkey RSA constructor again
net-ssh constructs the pkey without passing in any argument, which results
in a dummy key (it's up to the caller to populate the key in this case, which
is what net-ssh does). This commit handles this constructor case.

This is not documented in https://ruby-doc.org/stdlib-2.1.6/libdoc/openssl/rdoc/OpenSSL/PKey/RSA.html
but does work in vanilla Ruby (probably because this is a built-in C method).

Our keystores override still doesn't support constructing a key passing in the size (the
library we forked from didn't, and apparently we don't need to... yet).